requirements.txt
cassandra-driver==3.8.1
six==1.10.0
